router: move playground in internal package
